# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backend do Terminal Rodoviário",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "https://terminal-rodoviario-backend.herokuapp.com"
    },
    {
      "url": "http://localhost:8000"
    }
  ],
...
  "paths": {
...
    "/viacao": {
      "get": {
        "tags": ["Viação"],
        "summary": "Retorna uma lista com todas as viações ativas",
        "operationId": "fetchViacoes",
        "responses": {
          "200": {
            "description": "Caso a busca seja bem sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Viacao"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viacao/create": {
      "post": {
        "tags": ["Viação"],
        "summary": "Cria uma nova viação",
        "operationId": "createViacao",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "cnpj": {
                    "type": "string"
                  },
                  "imagem": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Caso a viação seja criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "cnpj": {
                      "type": "string"
                    },
                    "imagemURl": {
                      "type": "string"
                    },
                    "linhas": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "maxItems": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Caso não tenha sido enviada uma imagem ou caso o CPNJ enviado já tenha sido cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Imagem não enviada": {
                    "value": {
                      "message": "O campo imagem está vazio."
                    }
                  },
                  "CNPJ já cadastrado": {
                    "value": {
                      "message": "CNPJ já cadastrado."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viacao/delete/{viacaoId}": {
      "delete": {
        "tags": ["Viação"],
        "summary": "Deleta uma viação com base no id passado",
        "operationId": "deleteViacao",
        "parameters": [
          {
            "name": "viacaoId",
            "in": "path",
            "description": "Id da viação",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caso a viação seja excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viação deletada com sucesso."
                }
              }
            }
          },
          "400": {
            "description": "Caso o id passado não corresponda a uma viação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viação não existe."
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Viacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "imagemUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "linhas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Linha"
            }
          }
        }
      },
      "Linha": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pontoA": {
            "type": "string"
          },
          "pontoB": {
            "type": "string"
          },
          "valor": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "horarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Horario"
            }
          }
        }
      },
      "Horario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "horario": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Mensagem": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}